national_id_validator/requirements.txt

Django==4.0
django-cors-headers==4.4.0
django-filter==23.5
django-ratelimit==4.1.0
django-templated-mail==1.1.1
django-translations==1.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1

